Adjust the project structure in line with best practices.
