+++ LICENSE.md
@@ -1,4 +1,4 @@
-Copyright 2021 David Snowdon
+Copyright 2023 David Snowdon
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
