+++ setup.cfg
@@ -1,0 +1,49 @@
+
+ 
+
+[metadata]
+name = wsc_spot_poll
+url=https://bitbucket.org/wsctechnicalcommittee/wsc_spot_poll
+
+description = "Project for gathering GPS data from SPOT trackers into InfluxDB"
+long_description = file: README.md, LICENSE
+long_description_content_type = text/markdown
+
+author = "David Snowdon"
+author_email = "dcasnowdon+bitbucket@gmail.com"
+
+maintainer = "David Snowdon"
+maintainer_email = "dcasnowdon+bitbucket@gmail.com"
+
+license = "BSD 3-Clause License"
+license_files = file: LICENSE
+
+keywords = gps,spot
+
+classifiers =
+    License :: OSI Approved :: BSD License
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3
+    Framework :: Pytest
+    Intended Audience :: Developers
+    Topic :: Software Development :: Testing
+
+project_urls =
+    Source = https://bitbucket.org/wsctechnicalcommittee/wsc_spot_poll
+    Tracker = https://bitbucket.org/wsctechnicalcommittee/wsc_spot_poll
+    Documentation = https://bitbucket.org/wsctechnicalcommittee/wsc_spot_poll
+
+[options]
+packages = find:
+
+package_dir=
+    =src
+
+install_requires =
+    influxdb_client
+
+[options.packages.find]
+where=src
+
+[options.package_data]
+* = README.md
