+++ src/wsc_spot_poll/__init__.py
@@ -1,0 +1,6 @@
+hello="Hello!"
+
+print(hello)
+print("world")
+print("foo")
+print(hello)
