+++ src/wsc_spot_poll/__main__.py
